[fix] Added errorHandler to all controller routes
