+++ server/routes/modal.js
@@ -1,8 +1,9 @@
 import {Router} from "express";
 import modalContent from "../controllers/modal-content-controller";
+import errorHandler from "../utils/error-handler";
 
 const router = Router({mergeParams: true});
 
-router.get("/", modalContent);
+router.get("/", errorHandler(modalContent));
 
 export default router;
