+++ server/routes/payment.js
@@ -1,8 +1,9 @@
 import {Router} from "express";
 import payments from "../controllers/payments-controller";
+import errorHandler from "../utils/error-handler";
 
 const router = Router({mergeParams: true});
 
-router.post("/status/:paymentId", payments);
+router.post("/status/:paymentId", errorHandler(payments));
 
 export default router;
